requirements.txt
langgraph
langchain
langchain_openai
langchain_community
langchain_core
python-dotenv
langchain_pinecone
pinecone
jq